Fix client side warrior attack issue
When a client's own player (i.e. the warrior) attacked an enemy,
the player doesn't get added as an attacker to the attacked entity,
because the players attack logic is handled separately. When an entity
died the player however still got removed as an attacker with the
regular logic, resulting in an attacker being removed that wasn't there
to begin with.