Make game playable again
The host and port in the client configuration are set to null, which
lets the WebSocket URL building code to fill those in from the page's
location. This, in turn, matches LeetQuest's deployment model of a
single server serving everything (with, possibly, additional
infrastructure caching or originating static assets in front).